requirements.txt
Flask==2.2.5
python-docx==0.8.11
